feat: static table metadata access support
